share TestMain
