+++ rpc/tendermint/test/genesis.go
@@ -1,6 +1,6 @@
 package rpctest
 
-// priv keys generated deterministically eg rpc/tests/shared_test.go
+// priv keys generated deterministically eg rpc/tests/shared.go
 var defaultGenesis = `{
   "chain_id" : "tendermint_test",
   "accounts": [
